Setting version as 3.0.0
At this point I am not really sure we will call it 3.0.0 (TBD)

however I'm debugging differences between main and new-logging and I have to keep rebuilding when I switch folders as they are stored on maven repo. This will make debug scenarios a bit easier.

We can have the discussion on using 3.0 or 2.x later.