Revert "Remove slf4j bridge depedencies from libs", towards getting build back to consistent deps state and to more easily revert original addition leading to issue.
Th shading failing on rebuild issue isnt to do with competing service files from the JCL->SLF4J bridge and other logging libs, there really is only one, its just that the shade-plugin and/or the config being used means it behaves differently on rebuild, trying to repackage its own previous already-repackaged output on the 2nd build, and failing as it is then both transforming an already-shaded output and separately shading the original again and conflicting with itself. This 'differnt behaviour on rebuild' also happens on main, its just that none of the service files there actually get shaded but just copied unchanged, so it doenst trip over itself doing 2 actions on them. Configuring it to not transform the already-shaded file again stops the issue, but will look to go a different way per opening summary.

This reverts commit 26a37860699e2e4f99d28920bd3c1ab469cffd47.